feat: Bunch of improvements
